refactor(runtime): four more files hand their second jobs to named collaborators
CommandRules was four rule families riding one class; each now reads in
its own file — command_rules/admissibility (givens + lifecycle),
references (a reference points at something that exists), arithmetic
(MutationOp table, source resolution, increment/decrement), emission.
SagaInterpreter hands its three-tier correlation reading to
saga_interpreter/correlation.rb. EraGuard splits its pure half into
era_guard/shape_diff.rb — the signature/diff computation the Postgres
mint path calls too — keeping the file-boot check where it was.
Registry keeps the collections and hands the wiring gate to
registry/verification.rb. Every public name stays reachable exactly
where callers know it.
