+++ lib/hecksagain/runtime/registry.rb
@@ -1,8 +1,15 @@
+require_relative "registry/verification"
+
 module Hecksagain
   module Runtime
     class WiringError < StandardError; end
 
+    # The collections a boot gathers — bluebooks, hexagons, ports, adapters,
+    # worlds, the logs — and how a repository is resolved from them. The
+    # wiring gate lives in registry/verification.rb.
     class Registry
+      include Verification
+
       attr_reader :root, :bluebooks, :hecksagons, :ports, :adapters, :worlds, :event_log,
                   :reaction_log, :saga_log, :saga_instances, :translations
 
@@ -84,78 +91,6 @@ module Hecksagain
         false
       end
 
-      def verify!
-        verify_default_adapter!
-
-        @hecksagons.each_value do |hexagon|
-          hexagon.binds.each do |bind|
-            aggregate = bluebook(hexagon.domain)&.aggregate(bind.aggregate_name)
-            raise WiringError, "#{bind.aggregate} is bound but not declared in the bluebook" unless aggregate
-
-            check_verb(bind)
-
-            repository(hexagon.domain, aggregate)
-          end
-        end
-        self
-      end
-
-      def verify_default_adapter!
-        name = Ports::Persistence::DEFAULT_ADAPTER
-
-        check_verb(
-          Bluebook::IR::Bind.new(
-            aggregate: "(default)",
-            verb:      Ports::Persistence::VERB,
-            adapter:   name
-          )
-        )
-        adapter_class(name)
-        self
-      rescue WiringError => error
-        raise WiringError,
-              "the default persistence adapter (#{name}) is not usable, so an " \
-              "aggregate with no bind could not be given one: #{error.message}"
-      end
-
-      def check_verb(bind)
-        port = port_for(bind)
-        return if port.verb.to_s == bind.verb.to_s
-
-        raise WiringError,
-              "#{bind.adapter} implements the #{port.name} port (verb #{port.verb}) " \
-              "and cannot satisfy #{bind.verb}"
-      end
-
-
-      def check_settings(bind, settings)
-        adapter = @adapters[bind.adapter]
-        return unless adapter
-
-        declared = settings.keys - [:adapter]
-        unknown  = declared.reject { |field| adapter.declares?(field) }
-        return if unknown.empty?
-
-        raise WiringError,
-              "#{bind.adapter} does not declare #{unknown.map(&:inspect).join(', ')} — " \
-              "it declares #{adapter.all_fields.map(&:inspect).join(', ')}. " \
-              "Add the field to the adapter, or remove it from the world."
-      end
-
-      def port_for(bind)
-        adapter = @adapters[bind.adapter]
-        raise WiringError, "unknown adapter #{bind.adapter.inspect}" unless adapter
-
-        @ports[adapter.port] ||
-          raise(WiringError, "adapter #{bind.adapter} declares unknown port #{adapter.port.inspect}")
-      end
-
-      def adapter_class(name)
-        Adapters.const_get(name)
-      rescue NameError
-        raise WiringError, "no Ruby adapter implementation for #{name.inspect} " \
-                           "(expected Hecksagain::Adapters::#{name})"
-      end
     end
   end
 end
