+++ lib/hecksagain/runtime/registry/verification.rb
@@ -1,0 +1,85 @@
+module Hecksagain
+  module Runtime
+    class Registry
+      # The wiring gate: every bind names a declared aggregate, every
+      # adapter satisfies the verb its port declares, every world setting is
+      # a field the adapter admits, and the default adapter is usable at
+      # all. Included into Registry — `verify!` is what a boot calls after
+      # loading, and the smaller checks are also called piecemeal by the
+      # repository factory.
+      module Verification
+        def verify!
+          verify_default_adapter!
+
+          @hecksagons.each_value do |hexagon|
+            hexagon.binds.each do |bind|
+              aggregate = bluebook(hexagon.domain)&.aggregate(bind.aggregate_name)
+              raise WiringError, "#{bind.aggregate} is bound but not declared in the bluebook" unless aggregate
+
+              check_verb(bind)
+
+              repository(hexagon.domain, aggregate)
+            end
+          end
+          self
+        end
+
+        def verify_default_adapter!
+          name = Ports::Persistence::DEFAULT_ADAPTER
+
+          check_verb(
+            Bluebook::IR::Bind.new(
+              aggregate: "(default)",
+              verb:      Ports::Persistence::VERB,
+              adapter:   name
+            )
+          )
+          adapter_class(name)
+          self
+        rescue WiringError => error
+          raise WiringError,
+                "the default persistence adapter (#{name}) is not usable, so an " \
+                "aggregate with no bind could not be given one: #{error.message}"
+        end
+
+        def check_verb(bind)
+          port = port_for(bind)
+          return if port.verb.to_s == bind.verb.to_s
+
+          raise WiringError,
+                "#{bind.adapter} implements the #{port.name} port (verb #{port.verb}) " \
+                "and cannot satisfy #{bind.verb}"
+        end
+
+        def check_settings(bind, settings)
+          adapter = @adapters[bind.adapter]
+          return unless adapter
+
+          declared = settings.keys - [:adapter]
+          unknown  = declared.reject { |field| adapter.declares?(field) }
+          return if unknown.empty?
+
+          raise WiringError,
+                "#{bind.adapter} does not declare #{unknown.map(&:inspect).join(', ')} — " \
+                "it declares #{adapter.all_fields.map(&:inspect).join(', ')}. " \
+                "Add the field to the adapter, or remove it from the world."
+        end
+
+        def port_for(bind)
+          adapter = @adapters[bind.adapter]
+          raise WiringError, "unknown adapter #{bind.adapter.inspect}" unless adapter
+
+          @ports[adapter.port] ||
+            raise(WiringError, "adapter #{bind.adapter} declares unknown port #{adapter.port.inspect}")
+        end
+
+        def adapter_class(name)
+          Adapters.const_get(name)
+        rescue NameError
+          raise WiringError, "no Ruby adapter implementation for #{name.inspect} " \
+                             "(expected Hecksagain::Adapters::#{name})"
+        end
+      end
+    end
+  end
+end
